Seperate battery reader and display_information for update battery informations
