+++ riberry/battery/__init__.py
@@ -1,4 +1,5 @@
 # flake8: noqa
 
+from .common import decide_battery_i2c_bus_number
 from .pisugar import PisugarBatteryReader
 from .mp2760 import MP2760BatteryMonitor
